+++ hack/release.toml
@@ -16,10 +16,24 @@ preface = """\
 [notes]
 
     [notes.join]
-        title = "Join node type"
+        title = "Join Node Type"
         description = """\
 Node type `join` was renamed to `worker` for clarity. The old value is still accepted in the machine configuration but deprecated.
 `talosctl gen config` now generates `worker.yaml` instead of `join.yaml`.
+"""
+
+    [notes.bootkube]
+        title = "Support for Self-hosted Control Plane Dropped"
+        description = """\
+> **Note**: This item only applies to clusters bootstrapped with Talos <= 0.8.
+
+Talos 0.12 completely removes support for self-hosted Kubernetes control plane (bootkube-based).
+Talos 0.9 introduced support for Talos-managed control plane and provided migration path to convert self-hosted control plane
+to Talos-managed static pods.
+Automated and manual conversion process is available in Talos from 0.9.x to 0.11.x.
+For clusters bootstrapped with bootkube (Talos <= 0.8), please make sure control plane is converted to Talos-managed before
+before upgrading to Talos 0.12.
+Current control plane status can be checked with `talosctl get bootstrapstatus` before performing upgrade to Talos 0.12.
 """
