+++ internal/app/machined/pkg/runtime/v1alpha1/v1alpha1_sequencer.go
@@ -253,10 +253,6 @@ func (*Sequencer) Boot(r runtime.Runtime) []runtime.Phase {
 		r.State().Platform().Mode() != runtime.ModeContainer,
 		"bootloader",
 		UpdateBootloader,
-	).AppendWhen(
-		r.Config().Machine().Type() != machine.TypeWorker,
-		"checkControlPlaneStatus",
-		CheckControlPlaneStatus,
 	)
 
 	return phases
